IApp moved to App folder
