+++ TagCloud/App/IApp.cs
@@ -1,4 +1,4 @@
-namespace TagCloud
+namespace TagCloud.App
 {
     public interface IApp
     {
